+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
  * added ability to define an index for services in an injected service locator argument
  * made `ServiceLocator` implement `ServiceProviderInterface`
  * deprecated support for non-string default env() parameters
+ * added `%env(yaml:...)%` processor to parse YAML data
 
 4.2.0
 -----